python_server/requirements.txt
boto3~=1.34.76
pandas~=2.2.1
numpy~=1.26.4
scipy~=1.13.0
scikit-learn~=1.4.1.post1
logger~=1.4
pydantic~=2.6.4
fastapi~=0.110.1
SQLAlchemy~=2.0.29
python-dotenv~=1.0.1
uvicorn>=0.15.0,<0.16.0
mysqlclient~=2.2.4